requirements.txt
Flask
python-dotenv
mysqlclient
gunicorn
requests